MazeCellとMazeCellModelを分離
